requirements.txt
langchain==0.1.3
langchain-openai==0.0.3
python-dotenv==1.0.1